Organized the different user group views into their own seperate folders
