+++ app/components/navigationButton.tsx
@@ -1,7 +1,6 @@
 import { ButtonProps, Pressable, Text } from 'react-native';
 import { styles } from '../styles';
 
-
 export default function NavigationButton(props: ButtonProps): JSX.Element{
 	const {onPress, title} = props;
 	return (
