+++ app/views/NAVViews/NavHomeView.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Alert } from 'react-native';
 import { Text, View } from 'react-native';
-import NavigationButton from '../components/navigationButton';
-import { styles } from '../styles';
+import NavigationButton from '../../components/navigationButton';
+import { styles } from '../../styles';
 
 export default function NAVHomeView(){
 	return (
