+++ app/views/SymfoniViews/SymfoniHomeView.tsx
@@ -1,26 +1,21 @@
 import React from 'react';
 import { Alert } from 'react-native';
 import { Text, View } from 'react-native';
-import NavigationButton from '../components/navigationButton';
-import { styles } from '../styles';
+import NavigationButton from '../../components/navigationButton';
+import { styles } from '../../styles';
 
-export default function UserHomeView(){
+export default function SymfoniHomeView(){
 	return (
 		<View style={styles.container}>
-			<Text>This is the user home page.</Text>
+			<Text>This is the symfoni home page.</Text>
 
 			<NavigationButton
-				title='Recieve VC'
+				title='Create employment contract'
 				onPress={()=> Alert.alert('Test','Test')}
 			></NavigationButton>
 
 			<NavigationButton
-				title='List VCs'
-				onPress={()=> Alert.alert('Test','Test')}
-			></NavigationButton>
-
-			<NavigationButton
-				title='Send VP'
+				title='Create termination contract'
 				onPress={()=> Alert.alert('Test','Test')}
 			></NavigationButton>
